refactor(payment): address PR review feedback
